requirements.txt
Django==3.2
djangorestframework==3.12.4
django-filter~=2.4.0
djoser==2.2.3
Pillow
reportlab
isort==5.13.2